+++ caching/disable-lscache-notice.php
@@ -4,7 +4,6 @@
  * Description: Disables the admin notice Litespeeds LSCache generates about conflicting plugins installed
  * Type: snippet
  * Status: WIP
- */
 
 This is a WIP and not functioning yet.